requirements.txt
ddt==1.6.0
requests==2.28.2
xlrd==1.2.0
logbook==1.5.3
xtestrunner~=1.6.2
pyyaml~=6.0
pytest~=7.2.2
colorlog~=6.7.0
simplejson~=3.18.4